update for the `pruned_transducer_stateless7` for aishell and librispeech
